+++ clap_derive/src/derives/from_arg_matches.rs
@@ -152,7 +152,7 @@ fn gen_parsers(
         Ty::Option => {
             if attrs.is_enum() {
                 if let Some(subty) = subty_if_name(&field.ty, "Option") {
-                    parse = gen_arg_enum_parse(subty, &attrs);
+                    parse = gen_arg_enum_parse(subty, attrs);
                 }
             }
 
@@ -183,7 +183,7 @@ fn gen_parsers(
         Ty::Vec => {
             if attrs.is_enum() {
                 if let Some(subty) = subty_if_name(&field.ty, "Vec") {
-                    parse = gen_arg_enum_parse(subty, &attrs);
+                    parse = gen_arg_enum_parse(subty, attrs);
                 }
             }
 
@@ -204,7 +204,7 @@ fn gen_parsers(
 
         Ty::Other => {
             if attrs.is_enum() {
-                parse = gen_arg_enum_parse(&field.ty, &attrs);
+                parse = gen_arg_enum_parse(&field.ty, attrs);
             }
 
             quote_spanned! { ty.span()=>
@@ -269,7 +269,7 @@ pub fn gen_constructor(fields: &Punctuated<Field, Comma>, parent_attribute: &Att
                 Some(val) => quote_spanned!(kind.span()=> #field_name: (#val).into()),
             },
 
-            Kind::Arg(ty) => gen_parsers(&attrs, ty, field_name, field, None),
+            Kind::Arg(ty) | Kind::FromGlobal(ty) => gen_parsers(&attrs, ty, field_name, field, None),
         }
     });
 
@@ -349,7 +349,7 @@ pub fn gen_updater(
 
             Kind::Skip(_) => quote!(),
 
-            Kind::Arg(ty) => gen_parsers(&attrs, ty, field_name, field, Some(&access)),
+            Kind::Arg(ty) | Kind::FromGlobal(ty) => gen_parsers(&attrs, ty, field_name, field, Some(&access)),
         }
     });