+++ src/lib.rs
@@ -7,7 +7,7 @@
 #![doc(html_logo_url = "https://clap.rs/images/media/clap.png")]
 #![doc(html_root_url = "https://docs.rs/clap/3.0.0-beta.2")]
 #![cfg_attr(feature = "doc", doc(include = "../README.md"))]
-//! https://github.com/clap-rs/clap
+//! <https://github.com/clap-rs/clap>
 #![crate_type = "lib"]
 #![deny(
     missing_docs,
@@ -18,6 +18,8 @@
     trivial_numeric_casts
 )]
 #![forbid(unsafe_code)]
+// TODO: https://github.com/rust-lang/rust-clippy/issues/7290
+#![allow(clippy::single_component_path_imports)]
 
 #[cfg(not(feature = "std"))]
 compile_error!("`std` feature is currently required to build `clap`");