+++ tests/subcommands.rs
@@ -362,7 +362,12 @@ fn issue_1161_multiple_hyphen_hyphen() {
     let res = App::new("myprog")
         .arg(Arg::new("eff").short('f'))
         .arg(Arg::new("pea").short('p').takes_value(true))
-        .arg(Arg::new("slop").takes_value(true).multiple(true).last(true))
+        .arg(
+            Arg::new("slop")
+                .takes_value(true)
+                .multiple_values(true)
+                .last(true),
+        )
         .try_get_matches_from(vec![
             "-f",
             "-p=bob",
@@ -452,4 +457,23 @@ fn subcommand_after_argument_looks_like_help() {
         .get_matches_from(vec!["myprog", "helt", "test"]);
     assert_eq!(m.value_of("some_text"), Some("helt"));
     assert_eq!(m.subcommand().unwrap().0, "test");
+}
+
+#[test]
+fn issue_2494_subcommand_is_present() {
+    let app = App::new("opt")
+        .arg(Arg::new("global").long("global"))
+        .subcommand(App::new("global"));
+
+    let m = app.clone().get_matches_from(&["opt", "--global", "global"]);
+    assert_eq!(m.subcommand_name().unwrap(), "global");
+    assert!(m.is_present("global"));
+
+    let m = app.clone().get_matches_from(&["opt", "--global"]);
+    assert!(m.subcommand_name().is_none());
+    assert!(m.is_present("global"));
+
+    let m = app.get_matches_from(&["opt", "global"]);
+    assert_eq!(m.subcommand_name().unwrap(), "global");
+    assert!(!m.is_present("global"));
 }